fix(scope): stop enforcing knowledge-graph mention scopes as the worker's write boundary
`validated-direct` is the knowledge graph's directModules for the issue text,
kept only where the file exists. Existence does not make a mentioned file the
edit target, and the ledger shows what enforcing it as one costs (vela,
2026-09-02, /work/* runs):

  source            runs  done  parked  avg attempts  scope-violation issues
  (none)             117    20      18             5                      20
  validated-direct    43     1      15            15                      15
  drafted             38     6      17             7                       7

Every scope violation this morning was a validated-direct reservation that
was simply wrong: "session list pin" reserved to `pipeline/auth/superthread.py`
while the worker edited `session_store.py` and the sessions router; a CI
timing issue reserved to `auth/github.py` while the fix was in
`.github/workflows/`. One task was held to its wrong scope 46 times.

`enforcedFileScope` is now the single reading of which provenances bind:
`declared` and `drafted`. `inferred` and `validated-direct` are advisory —
they still feed admission conflict checks, where over-caution only costs
ordering. The worker prompt, the post-run guard and the publication fence all
call it, so they cannot disagree about what is enforced.

`summarizeStageResult` moves to `stageSummary.ts` unchanged: pairPipeline.ts
was already past the 1500-line pre-commit cap before this change.
